# sms/service/server_service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for VCS System Management API.",
        "title": "VCS System Management API",
        "contact": {
            "name": "Vu Dang Duc Minh",
            "url": "https://github.com/vudangducminh",
            "email": "vudangducminh@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8801",
    "basePath": "/api/v1",
    "paths": {
...
        "/server/view_servers": {
            "get": {
                "description": "View server details with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "View server details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token for authentication",
                        "name": "jwt",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Order of results, either 'asc' or 'desc'. If not provided or using the wrong order format, the default order is ascending",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by server_name, ipv4, or status. If not provided or using the wrong filter format, the default filter is server_name",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Substring to search in server_name, ipv4, or status",
                        "name": "string",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/entities.ViewServerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/entities.ViewServerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/entities.AuthErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve server details",
                        "schema": {
                            "$ref": "#/definitions/entities.ViewServerInternalServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.AuthErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Authentication failed"
                }
            }
        },
...
        "entities.Server": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Unique identifier for the server",
                    "type": "string"
                },
                "created_time": {
                    "description": "Unix timestamp",
                    "type": "integer"
                },
                "ipv4": {
                    "type": "string"
                },
                "last_updated_time": {
                    "description": "Unix timestamp",
                    "type": "integer"
                },
                "server_name": {
                    "type": "string"
                },
                "status": {
                    "description": "e.g., \"active\", \"inactive\", \"maintenance\"",
                    "type": "string"
                },
                "uptime": {
                    "description": "e.g., \"3666\" for 1 hour 1 minute and 6 seconds",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "entities.ViewServerBadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
        "entities.ViewServerInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to retrieve server details"
                }
            }
        },
        "entities.ViewServerSuccessResponse": {
            "type": "object",
            "properties": {
                "servers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Server"
                    }
                }
            }
        }
    },
...
}